ENG-4079: #673: Replace intent type with intent type set
Summary:
Currently we allow only one type for each intent/lock being held by a transaction/operation, and we
have 6 of these types (for each combination of weak/strong on the one hand, and serializable read,
serializable write, and snapshot write on the other hand).  Because of this, we have a large (6x6)
manually constructed conflict matrix and a complex function that combines intent types to find a new
intent type that conflicts with the same set of other intent types as all of its input intent types
together, or a superset thereof.

Instead of that we could allow an lock/intent to have multiple types simultaneously.  So only 4
types of intents would be enough: weak read, strong read, weak write, strong write.

As a reminder, "weak" intents are intents that are acquired on a parent document of a document being
read or written, and "strong" intents are acquired on a subdocument that is being read in its
entirety, fully overwritten, or completely deleted.

Here is new the small and simple conflict matrix for these intent types:

```
    | WR | SR | WW | SW
------------------------
 WR | .    .    .    X
 SR | .    .    X    X
 WW | .    X    .    .
 SW | X    X    .    .
```

Our old snapshot write intent type could be represented as a combination of read and write intents
in the new scheme, because it has identical conflict properties.

This simplification gives us ability to use 64 bits for LockState, instead of 128 bits.

And the most important advantage of this approach is that when a single write request contains
different intent types for the same key, we can now combine them with the minimal possible number of
conflicts.

Test Plan: Jenkins

Reviewers: timur, robert, mikhail

Reviewed By: mikhail

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D5907

+++ src/yb/docdb/shared_lock_manager_fwd.h
@@ -22,7 +22,7 @@ namespace yb {
 namespace docdb {
 
 class LockBatch;
-typedef unsigned __int128 LockState;
+typedef uint64_t LockState;
 
 class SharedLockManager;
